Move MAME audio device emulations to `src/hardware/audio/`
