+++ src/hardware/audio/private/mame/emu.h
