+++ src/hardware/audio/private/mame/saa1099.h
