+++ src/hardware/audio/private/mame/sn76496.h
